+++ src/ts/AdUnits/VPAIDAdUnit.ts
@@ -1,26 +1,27 @@
 import { AbstractAdUnit, IAdUnitParameters } from 'AdUnits/AbstractAdUnit';
-import { VPAIDCampaign } from 'Models/VPAID/VPAIDCampaign';
-import { NativeBridge } from 'Native/NativeBridge';
-import { VPAID } from 'Views/VPAID';
+import { AdUnitContainerSystemMessage, IAdUnitContainerListener } from 'AdUnits/Containers/AdUnitContainer';
 import { FinishState } from 'Constants/FinishState';
 import { Platform } from 'Constants/Platform';
-import { OperativeEventManager } from 'Managers/OperativeEventManager';
-import { ThirdPartyEventManager } from 'Managers/ThirdPartyEventManager';
-import { Timer } from 'Utilities/Timer';
-import { Diagnostics } from 'Utilities/Diagnostics';
 import { DiagnosticError } from 'Errors/DiagnosticError';
-import { VPAIDEndScreen } from 'Views/VPAIDEndScreen';
-import { Closer } from 'Views/Closer';
+import { ThirdPartyEventManager } from 'Managers/ThirdPartyEventManager';
 import { DeviceInfo } from 'Models/DeviceInfo';
 import { Placement } from 'Models/Placement';
-import { IObserver2 } from 'Utilities/IObserver';
+import { VPAIDCampaign } from 'Models/VPAID/VPAIDCampaign';
 import { WKAudiovisualMediaTypes } from 'Native/Api/WebPlayer';
-import { AdUnitContainerSystemMessage, IAdUnitContainerListener } from 'AdUnits/Containers/AdUnitContainer';
+import { NativeBridge } from 'Native/NativeBridge';
+import { Diagnostics } from 'Utilities/Diagnostics';
+import { IObserver2 } from 'Utilities/IObserver';
+import { Timer } from 'Utilities/Timer';
+import { AbstractPrivacy } from 'Views/AbstractPrivacy';
+import { Closer } from 'Views/Closer';
+import { VPAID } from 'Views/VPAID';
+import { VPAIDEndScreen } from 'Views/VPAIDEndScreen';
 
 export interface IVPAIDAdUnitParameters extends IAdUnitParameters<VPAIDCampaign> {
     vpaid: VPAID;
     closer: Closer;
     endScreen?: VPAIDEndScreen;
+    privacy: AbstractPrivacy;
 }
 
 export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListener {
@@ -32,7 +33,6 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
     private static _adLoadTimeout: number = 10 * 1000;
     private _closer: Closer;
     private _placement: Placement;
-    private _operativeEventManager: OperativeEventManager;
     private _thirdPartyEventManager: ThirdPartyEventManager;
     private _view: VPAID;
     private _vpaidCampaign: VPAIDCampaign;
@@ -40,12 +40,12 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
     private _options: any;
     private _deviceInfo: DeviceInfo;
     private _urlLoadingObserver: IObserver2<string, string>;
+    private _privacyShowing = false;
 
     constructor(nativeBridge: NativeBridge, parameters: IVPAIDAdUnitParameters) {
         super(nativeBridge, parameters);
 
         this._vpaidCampaign = parameters.campaign;
-        this._operativeEventManager = parameters.operativeEventManager;
         this._thirdPartyEventManager = parameters.thirdPartyEventManager;
         this._options = parameters.options;
         this._view = parameters.vpaid;
@@ -55,6 +55,7 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
         this._timer = new Timer(() => this.onAdUnitNotLoaded(), VPAIDAdUnit._adLoadTimeout);
 
         this._closer.render();
+        this._closer.choosePrivacyShown();
     }
 
     public show(): Promise<void> {
@@ -62,6 +63,7 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
 
         return this.setupWebPlayer().then(() => {
             this._urlLoadingObserver = this._nativeBridge.WebPlayer.shouldOverrideUrlLoading.subscribe((url, method) => this.onUrlLoad(url));
+            this.setupPrivacyObservers();
             return this._container.open(this, ['webplayer', 'webview'], false, this._forceOrientation, false, false, true, false, this._options).then(() => {
                 this.onStart.trigger();
             });
@@ -125,6 +127,7 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
 
     public onContainerShow(): void {
         this.setShowing(true);
+        this.onContainerForeground();
     }
 
     public onContainerDestroy(): void {
@@ -137,15 +140,32 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
 
     public onContainerForeground(): void {
         this.showCloser();
-        if (this._view.isLoaded()) {
+        if (this._view.isLoaded() && !this._privacyShowing) {
             this._view.resumeAd();
-        } else {
+        } else if (!this._view.isLoaded()) {
             this._view.loadWebPlayer();
+        } else {
+            // Popup will resume video
         }
     }
 
     public onContainerSystemMessage(message: AdUnitContainerSystemMessage): void {
         // EMPTY
+    }
+
+    private setupPrivacyObservers(): void {
+        if (this._closer.onPrivacyClosed) {
+            this._closer.onPrivacyClosed.subscribe(() => {
+                this._view.resumeAd();
+                this._privacyShowing = false;
+            });
+        }
+        if (this._closer.onPrivacyOpened) {
+            this._closer.onPrivacyOpened.subscribe(() => {
+                this._view.pauseAd();
+                this._privacyShowing = true;
+            });
+        }
     }
 
     private setupWebPlayer(): Promise<any> {
@@ -214,21 +234,14 @@ export class VPAIDAdUnit extends AbstractAdUnit implements IAdUnitContainerListe
         this._container.removeEventHandler(this);
     }
 
-    private showView() {
-        this._view.show();
-    }
-
     private hideView() {
+        this._closer.hide();
         this._view.hide();
     }
 
     private showCloser() {
         return Promise.all([this._deviceInfo.getScreenWidth(), this._deviceInfo.getScreenHeight()]).then(([width, height]) => {
-            const left = Math.floor(width * 0.85);
-            const top = 0;
-            const viewWidth = Math.floor(width * 0.15);
-            const viewHeight = viewWidth;
-            return this._container.setViewFrame('webview', left, top, viewWidth, viewHeight).then(() => {
+            return this._container.setViewFrame('webview', 0, 0, width, height).then(() => {
                 if (!this._closer.container().parentNode) {
                     document.body.appendChild(this._closer.container());
                 }