+++ src/ts/EventHandlers/PromoEventHandler.ts
@@ -3,11 +3,12 @@ import { PurchasingUtilities, IPromoPayload, IPromoRequest } from 'Utilities/Pur
 import { Configuration } from 'Models/Configuration';
 import { FinishState } from 'Constants/FinishState';
 import { Placement } from 'Models/Placement';
-import { OperativeEventManager } from 'Managers/OperativeEventManager';
+import { ABGroup } from 'Models/ABGroup';
+import { GdprManager, GDPREventAction } from 'Managers/GdprManager';
 
 export class PromoEventHandler {
 
-    public static onClose(adUnit: PromoAdUnit, gamerToken: string, gameId: string, abGroup: number, purchaseTrackingUrls: string[], isOptOutEnabled: boolean): void {
+    public static onClose(adUnit: PromoAdUnit, gamerToken: string, gameId: string, abGroup: ABGroup, purchaseTrackingUrls: string[], isOptOutEnabled: boolean): void {
         adUnit.setFinishState(FinishState.SKIPPED);
         adUnit.hide();
         const iapPayload: IPromoPayload = {
@@ -15,7 +16,7 @@ export class PromoEventHandler {
             trackingOptOut: isOptOutEnabled,
             iapPromo: true,
             gameId: gameId + '|' + gamerToken,
-            abGroup: abGroup,
+            abGroup: abGroup.toNumber(),
             request: IPromoRequest.CLOSE,
             purchaseTrackingUrls: purchaseTrackingUrls,
         };
@@ -35,10 +36,10 @@ export class PromoEventHandler {
         PurchasingUtilities.sendPromoPayload(JSON.stringify(iapPayload));
     }
 
-    public static onGDPRPopupSkipped(configuration: Configuration, operativeEventManager: OperativeEventManager): void {
+    public static onGDPRPopupSkipped(configuration: Configuration, gdprManager: GdprManager): void {
         if (!configuration.isOptOutRecorded()) {
             configuration.setOptOutRecorded(true);
-            operativeEventManager.sendGDPREvent('skip');
+            gdprManager.sendGDPREvent(GDPREventAction.SKIP);
         }
     }
 }