+++ src/ts/Test/Unit/EventHandlers/PerformanceOverlayEventHandlerTest.ts
@@ -72,7 +72,7 @@ describe('PerformanceOverlayEventHandlerTest', () => {
 
         const privacy = new Privacy(nativeBridge, configuration.isCoppaCompliant());
         endScreen = new PerformanceEndScreen(nativeBridge, campaign, deviceInfo.getLanguage(), clientInfo.getGameId(), privacy, true);
-        overlay = new Overlay(nativeBridge, false, 'en', clientInfo.getGameId());
+        overlay = new Overlay(nativeBridge, false, 'en', clientInfo.getGameId(), privacy, false);
         const gdprManager = sinon.createStubInstance(GdprManager);
 
         performanceAdUnitParameters = {