+++ src/ts/Test/Unit/EventHandlers/PerformanceVideoEventHandlersTest.ts
@@ -67,7 +67,7 @@ describe('PerformanceVideoEventHandlersTest', () => {
 
         const privacy = new Privacy(nativeBridge, configuration.isCoppaCompliant());
         endScreen = new PerformanceEndScreen(nativeBridge, campaign, deviceInfo.getLanguage(), clientInfo.getGameId(), privacy, false);
-        overlay = new Overlay(nativeBridge, false, 'en', clientInfo.getGameId());
+        overlay = new Overlay(nativeBridge, false, 'en', clientInfo.getGameId(), privacy, false);
         const gdprManager = sinon.createStubInstance(GdprManager);
 
         performanceAdUnitParameters = {