+++ src/ts/Utilities/PurchasingUtilities.ts
@@ -69,10 +69,10 @@ export class PurchasingUtilities {
                 this._nativeBridge.Purchasing.onGetPromoCatalog.unsubscribe(observer);
                 try {
                     this._catalog = new PurchasingCatalog(JSON.parse(promoCatalogJSON));
+                    resolve();
                 } catch(err) {
                     reject(this.logIssue('catalog_json_parse_failure', 'Promo catalog JSON failed to parse'));
                 }
-                resolve();
             });
             this._nativeBridge.Purchasing.getPromoCatalog().catch((e) => {
                 this._nativeBridge.Purchasing.onGetPromoCatalog.unsubscribe(observer);
@@ -145,12 +145,13 @@ export class PurchasingUtilities {
                 this._nativeBridge.Purchasing.onInitialize.unsubscribe(observer);
                 if (isReady !== 'True') {
                     reject(this.logIssue('promo_not_ready', 'IAP Promo was not ready'));
+                } else {
+                    resolve();
                 }
-                resolve();
             });
             this._nativeBridge.Purchasing.initializePurchasing().catch(() => {
                 this._nativeBridge.Purchasing.onInitialize.unsubscribe(observer);
-                reject(this.logIssue('purchase_initilization_failed', 'Purchase initialization failed'));
+                reject(this.logIssue('purchase_initialization_failed', 'Purchase initialization failed'));
             });
         });
     }
@@ -161,8 +162,9 @@ export class PurchasingUtilities {
                 this._nativeBridge.Purchasing.onGetPromoVersion.unsubscribe(promoVersionObserver);
                 if(!this.isPromoVersionSupported(promoVersion)) {
                     reject(this.logIssue('promo_version_not_supported', 'Promo version not supported'));
+                } else {
+                    resolve();
                 }
-                resolve();
             });
             this._nativeBridge.Purchasing.getPromoVersion().catch(() => {
                 this._nativeBridge.Purchasing.onGetPromoVersion.unsubscribe(promoVersionObserver);
@@ -179,8 +181,9 @@ export class PurchasingUtilities {
                         this._isInitialized = true;
                     }
                     resolve();
+                } else {
+                    reject(this.logIssue('purchase_command_attempt_failed', `Purchase command attempt failed with command ${isCommandSuccessful}`));
                 }
-                reject(this.logIssue('purchase_command_attempt_failed', 'Purchase command attempt failed'));
             });
             this._nativeBridge.Purchasing.initiatePurchasingCommand(iapPayload).catch(() => {
                 this._nativeBridge.Purchasing.onCommandResult.unsubscribe(observer);
@@ -192,7 +195,7 @@ export class PurchasingUtilities {
     // Returns true if version is 1.16.0 or newer
     private static isPromoVersionSupported(version: string): boolean {
         const promoVersionSplit = version.split('.', 2);
-        if (promoVersionSplit[0].length > 0 && promoVersionSplit[0].length > 0) {
+        if (promoVersionSplit.length > 1) {
             return ((parseInt(promoVersionSplit[0], 10) >= 2) || ((parseInt(promoVersionSplit[0], 10) >= 1 && parseInt(promoVersionSplit[1], 10) >= 16)));
         }
         return false;
@@ -201,7 +204,7 @@ export class PurchasingUtilities {
     private static loadInitializationPayloads(): IPromoPayload {
         return <IPromoPayload>{
             iapPromo: true,
-            abGroup: this._configuration.getAbGroup(),
+            abGroup: this._configuration.getAbGroup().toNumber(),
             gameId: this._clientInfo.getGameId() + '|' + this._configuration.getToken(),
             trackingOptOut: this._configuration.isOptOutEnabled(),
             gamerToken: this._configuration.getToken(),