+++ src/ts/Views/AbstractVideoOverlay.ts
@@ -7,6 +7,7 @@ export interface IOverlayHandler {
     onOverlayPauseForTesting(paused: boolean): void;
     onOverlayCallButton(): void;
     onOverlayClose(): void;
+    onGDPRPopupSkipped(): void;
 }
 
 export abstract class AbstractVideoOverlay extends View<IOverlayHandler> {
@@ -17,7 +18,6 @@ export abstract class AbstractVideoOverlay extends View<IOverlayHandler> {
 
     protected static AutoSkip: boolean = false;
 
-    protected _abGroup: number;
     protected _skipDuration: number;
     protected _skipRemaining: number;
 
@@ -26,10 +26,10 @@ export abstract class AbstractVideoOverlay extends View<IOverlayHandler> {
     protected _muted: boolean;
 
     protected _fadeEnabled: boolean = true;
+    protected _isPrivacyShowing: boolean = false;
 
-    constructor(nativeBridge: NativeBridge, containerId: string, muted: boolean, abGroup: number = 0) {
+    constructor(nativeBridge: NativeBridge, containerId: string, muted: boolean) {
         super(nativeBridge, containerId);
-        this._abGroup = abGroup;
         this._muted = muted;
     }
 
@@ -49,6 +49,10 @@ export abstract class AbstractVideoOverlay extends View<IOverlayHandler> {
         if(this._fadeEnabled !== value) {
             this._fadeEnabled = value;
         }
+    }
+
+    public isPrivacyShowing(): boolean {
+        return this._isPrivacyShowing;
     }
 
     public abstract setSpinnerEnabled(value: boolean): void;