+++ src/ts/Views/PlayableMRAID.ts
@@ -13,6 +13,7 @@ import { IMRAIDViewHandler, MRAIDView } from 'Views/MRAIDView';
 import { SdkStats } from 'Utilities/SdkStats';
 import { AbstractPrivacy } from 'Views/AbstractPrivacy';
 import { CustomFeatures } from 'Utilities/CustomFeatures';
+import { PlayableEndScreenHideDelayDisabledAbTest } from 'Models/ABGroup';
 
 export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
 
@@ -23,6 +24,9 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
     private _closeElement: HTMLElement;
     private _loadingScreen: HTMLElement;
     private _iframe: HTMLIFrameElement;
+    private _gdprBanner: HTMLElement;
+    private _privacyButton: HTMLElement;
+
     private _iframeLoaded = false;
 
     private _messageListener: any;
@@ -42,8 +46,8 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
 
     private _configuration: any;
 
-    constructor(nativeBridge: NativeBridge, placement: Placement, campaign: MRAIDCampaign, language: string, privacy: AbstractPrivacy) {
-        super(nativeBridge, 'playable-mraid', placement, campaign, privacy);
+    constructor(nativeBridge: NativeBridge, placement: Placement, campaign: MRAIDCampaign, language: string, privacy: AbstractPrivacy, showGDPRBanner: boolean) {
+        super(nativeBridge, 'playable-mraid', placement, campaign, privacy, showGDPRBanner);
 
         this._placement = placement;
         this._campaign = campaign;
@@ -80,6 +84,14 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
                 event: 'click',
                 listener: (event: Event) => this.onPrivacyEvent(event),
                 selector: '.privacy-button'
+            },
+            {
+                event: 'click',
+                listener: (event: Event) => {
+                    this.onGDPRPopupEvent(event);
+                    this.choosePrivacyShown();
+                },
+                selector: '.gdpr-link'
             }
         ];
     }
@@ -91,6 +103,8 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
         this._loadingScreen = <HTMLElement>this._container.querySelector('.loading-screen');
 
         const iframe: any = this._iframe = <HTMLIFrameElement>this._container.querySelector('#mraid-iframe');
+        this._gdprBanner = <HTMLElement>this._container.querySelector('.gdpr-pop-up');
+        this._privacyButton = <HTMLElement>this._container.querySelector('.privacy-button');
 
         let container = MRAIDContainer;
         const playableConfiguration = this._campaign.getPlayableConfiguration();
@@ -115,6 +129,8 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
 
         this._messageListener = (event: MessageEvent) => this.onMessage(event);
         window.addEventListener('message', this._messageListener, false);
+
+        this.choosePrivacyShown();
     }
 
     public show(): void {
@@ -165,6 +181,18 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
                     this._backgroundTime += Date.now() - this._backgroundTimestamp;
                 }
             }
+        }
+    }
+
+    protected choosePrivacyShown(): void {
+        if (this._showGDPRBanner && !this._gdprPopupClicked) {
+            this._gdprBanner.style.visibility = 'visible';
+            this._privacyButton.style.pointerEvents = '1';
+            this._privacyButton.style.visibility = 'hidden';
+        } else {
+            this._privacyButton.style.visibility = 'visible';
+            this._gdprBanner.style.pointerEvents = '1';
+            this._gdprBanner.style.visibility = 'hidden';
         }
     }
 
@@ -272,7 +300,7 @@ export class PlayableMRAID extends MRAIDView<IMRAIDViewHandler> {
             }, false);
         });
 
-        if (CustomFeatures.isPlayableEndScreenHideDelayDisabled(this._campaign.getAbGroup())) {
+        if (PlayableEndScreenHideDelayDisabledAbTest.isValid(this._campaign.getAbGroup())) {
             this._loadingScreen.classList.add('disable-delay');
         }